fix guild syntax to not be registered as a mode
Changed guild syntax so that it isn't registered as a mode so that you can't do +pmode guild and make it throw errors. Also so that you can't try to use stats_multiple with guild by doing +mode guild and then listing multiple users causing it to throw more errors. I also added an invalid mode to logging, because if it was an invalid mode, it would say for example: "Party Mode: FatDubs -->         ", and it would just be empty.